Post to Mattermost all scheduled command reports.

+++ tests/erp/export/test_export_integration.py
@@ -34,9 +34,7 @@ def test_csv_creation(db):
 
 def test_export_command(mocker, db, settings):
     settings.DATAGOUV_API_KEY = "fake"  # To pass the check before uploading
-    mocker.patch(
-        "requests.post",
-    )
+    mocker.patch("requests.post")
 
     management.call_command("export_to_datagouv")
     assert os.path.isfile("acceslibre.csv")
@@ -48,10 +46,16 @@ def test_export_failure(mocker, db, settings):
     settings.DATAGOUV_API_KEY = "fake"  # To pass the check before uploading
     mocker.patch(
         "requests.post",
-        side_effect=requests.exceptions.HTTPError("Error"),
+        side_effect=requests.RequestException("KO"),
     )
-    with pytest.raises(RuntimeError) as err:
-        management.call_command("export_to_datagouv")
+    from erp.management.commands.export_to_datagouv import logger
 
-    os.unlink("acceslibre.csv")
-    assert "Erreur lors de l'upload" in str(err.value)
+    spy = mocker.spy(logger, "error")
+
+    management.call_command("export_to_datagouv")
+    if os.path.isfile("acceslibre.csv"):
+        os.unlink("acceslibre.csv")
+
+    spy.assert_called_once_with(
+        "Impossible de publier le dataset: Erreur lors de l'upload: KO"
+    )
